+++ src/app/mypage/components/Gnb/index.tsx
