+++ src/app/mypage/components/Tab/index.ts
@@ -1,0 +1,2 @@
+export { default as Tab } from './Tab';
+export { default as TabGroup } from './TabGroup';
