+++ src/app/mypage/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Gnb } from './Gnb';
+export { default as Header } from './Header';
+export * from './Tab';
