+++ src/app/mypage/place/bookmarks/components/index.ts
@@ -1,0 +1,1 @@
+export { default as BookmarkItem } from './BookmarkItem';
