+++ src/app/mypage/place/reviews/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewItem } from './ReviewItem';
