+++ src/app/mypage/record/timeline/components/RecordItem/Dot.tsx
