+++ src/app/mypage/record/timeline/components/Weekly/index.tsx
