+++ src/app/mypage/record/timeline/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Monthly } from './Monthly';
+export { default as Weekly } from './Weekly';
+export { default as RecordItem } from './RecordItem';
