--- src/components/mypage/index.tsx
@@ -1,10 +1,0 @@
-export { default as Tab } from './Tab';
-export { default as TabGroup } from './TabGroup';
-export { default as Gnb } from './Gnb';
-export { default as Header } from './Header';
-export { default as BookmarkItem } from './BookmarkItem';
-export { default as ReviewItem } from './ReviewItem';
-export { default as RecordItem } from './RecordItem';
-export { default as Dot } from './Dot';
-export { default as Monthly } from './Monthly';
-export { default as Weekly } from './Weekly';
